Rename layers -> get-layers
