+++ packages/alfa-rules/src/common/dom/get-colors/get-layers.ts
